requirements.txt
confluent_kafka
psycopg2-binary
python-dotenv
pytest